backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.4
httpx==0.27.0
python-dotenv==1.0.1
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
google-cloud-aiplatform==1.56.0
google-generativeai==0.5.4
antigravity==0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
